✨ amp-selector added to kpi-list
the desktop design for kpi-list calls for wrapping any additional elements into a new row instead of rendering the amp-carousel. Amp-layout currently seems to not work with this feature, however a workable solution can be implemented with amp-selector; this example was taken from ampproject.org. A better solution should be implemented if found in the future  #49