🎉 new divider atom for lists
Some lists throughout the design require a divider between items. This divider is a simple atom with only a scss file that renders the divider when given to a span. Whisper2 color is also defined in variables, the name generator used labeled this color whisper and since whisper is already defined in variables the new color is labelled as whisper2 #54