✨ kpi items now take up their parent container
In the old structure Kpi items had a fixed width and height for desktop along with it's parent grid container also defining the height and width of each grid item. Now the KPI takes up the full width and height of the specified dimensions given to the grid-columns. The idea behind this is instead of setting the height and width for two items we only need to set the height and width for the grid columns and rows #54