✨ removed AMP carousel
AMP carousel required at least a fixed height which was not working well with the design of the KPI molecule. We've removed the AMP carousel in favor of a simple flex container with overflow: scroll to acheive the behavior we'd like without scarficing the overall design of the KPI molecule.   #49