🐛 grid layout fix for float image
components have grid layout properties and should be wrapped in a grid container. This was broken in the new grav alternative and is now fixed. All .col classes were removed and only those that are needed should be defined as this could cause us to hit out css 50kb limit. Class name .col-full was added for components that should take up the entire parent grid layout.  #63