♻️ refactored default dropdown stylings
Moved animations from the default dropdown.scss file to the stage-destionation file. This allows for the drowdown style file to be used in the docs sidebar without importing the animations that are used in other versions of the stage dropdown. This will cut down on the css file size for documentation page(s) #102