Feature/fix invalid story page (#1801)
* 🍱 updated low quality story teaser assets

single video and single image assets were of a lower quality and should be updated  #1764

* 🐛 amp story best practices was invalid amp

amp-video does not support intrinsic layout property, all layouts for amp-video are chagned to responsive and include a container with an inline style max-width based on their width property. This makes it easier to maintain incase any future videos require a different width size.  #1796