Tweak tools page styling and structure (#2316)
* :art: Auto-remove whitespace

#2234

* :bug: Increase bounding-box of tool-format-filter

Stops box-shadows of filter-bubbles cutting the edge of the bounding-box #2234

* :art: Better solve negative margin hack for filter-caption

Set grid-row-gap of filter-container to zero instead #2234

* :art: Move bgColor value to amp-image parent

This is needed one level higher, because of the padding for logos it will get #2234

* :art: Add fit value also to parent element

Since we use the 'contain' class only for logos, we can use the same class for adding padding to all images with that setting #2234

* :bug: Remove fixed height from sidebar

Fixes that sidebar is scrolling away too early if there is not enough content to fill all the height #2195

* :lipstick: Reduce padding-bottom of toc to fit with sidebar

#2195

* :fire: Remove deprecated grid-toggle

The grid-toggle isn't really up to date and don't fit anywhere correctly anymore. I originally build it to help developers finding bugs, but I don't think somebody is still using it. If someone will miss it, we can bringt it back and fix it.

* :art: Move two teaser to new tools category

TBD in which order the new category should appear #2234

* :art: Improve markup and cleanup css of tool-sections

Put section-title and teaser-grid together and give title a class #2234

* :sparkles: Add helper elements to tool-section-title

A little css trick to have control over empty tool-sections. Right now the opacity of the section-title reduces if the section has no visible teasers inside. TBD if another effect could be better #2234

* :art: Auto-remove whitespace

#2234

* :bug: Fix wrong variable

#2234

* :bento: Clean old map

Remove dots and give proper size #2094

* :sparkles: Add roadshow-map-helper-grid

This is useful for determining the desired coordinates for dots/cities that are upcoming in the future. It is only visible in dev environment #2094

* :wheelchair: Add tabindex and role to helper-grid items

#2094

* Revert ":wheelchair: Add tabindex and role to helper-grid items"

This reverts commit b944b03c8421f4b45f05dd6adfaffd076fba08c1.

* Revert ":sparkles: Add roadshow-map-helper-grid"

This reverts commit c9d3621ad2ff8969600335f9d890767a8eb0adce.

* Revert ":bento: Clean old map"

This reverts commit 316d60b30f4c9e2e7f8d0728c91cb8f36aace841.

* :recycle: Further improve markup and styling of tools page

#2234

* :speech_balloon: Rename category to more general name

To prevent doublings in the name I also changed the type of the cards to just 'Email Service'. #2234 #2316