Search meta tag page-locale has untranslated locales on default page
Currently we search for locale 'en' and the specific locale when we are
not in the 'en' language because the untranslated pages are not in the
google index.
With this change the 'en' documents also have the language in the meta
tags that are not translated, so we can simply search for the language
after the index was updated.