Add keyboard-select-mode
I've added the `keyboard-select-mode` attribute to each of the three examples on this sample page. 
It's worth noting that I've passed the `focus` option (versus `select` or `none`). There was no stipulation as to which option these examples ought to employ, so I went with what I thought was the most approachable option.  Please let me know if you would like me to make any changes.