Implement isCacheUrl check and rule for status banner
The check is implemented in the linter backend api as isCacheUrl.
The frontend linter check component translates this into isNotCacheURL to have all successful checks "true".
The difference is mostly due to the fact that we use the prod api on stage and with this setup pixi will not show the "cache url" status when the new flag is not in the linter api response.