Remove "Animate & Transition" from stories documentation
Currently, this expand/collapse control in the left-hand sidebar is non-functional, as there are no documentation pages in this directory that are actually marked as applying to "stories". Removing this for now, as presumably in future the existing documentation pages will include content specific to `<amp-story-animation>`, but for now this just makes the documentation look broken otherwise.