Restructure "optimized images" section
Currently, this page appears badly broken in https://amp.dev/documentation/guides-and-tutorials/develop/style_and_layout/responsive_design/?format=stories

This is due to the fact that the markdown processor does not seem to understand that after the `[sourcecode]` block the unordered list is supposed to continue, and instead closes all currently open containers and starts a fresh unordered list. Avoiding the problem by instead using subheadings and a few tweaks to the prose.