Attempt to fix missing sidebar on "Make your Web Stories discoverable
Not sure if this fix will do have the full effect, but: currently https://amp.dev/documentation/guides-and-tutorials/optimize-and-measure/discovery_web_stories/?format=stories seems to have a problem/lack the left-hand sidebar. This *may* be due to the unorthodox frontmatter. Normalised this to use the same properties, and order, of another sibling markdown document that *does* work correctly, hoping it's just related to that...