Fix component auto import
This fixes the problem that the validation rules include experimental
component versions causing bento components to be auto imported.

Instead of generating the component versions from validation rules, we
extract the latest version from the component docs.

This PR also removes the buildComponentVersion task as it is no longer
needed.