♿ TetraLogical: AMP Stories accessibility review / documentation (#5970)
* Typo

* Add bullet forr alt text, remove stray hanging spaces at line endings

* Typos

* Typos

* Typo

* Add color contrast note to background color section

* Re-title "desktop experience" section, correct link, expand explanation.

* "an Web Story" > "a Web Story"

* Remove "Animate & Transition" from stories documentation

Currently, this expand/collapse control in the left-hand sidebar is non-functional, as there are no documentation pages in this directory that are actually marked as applying to "stories". Removing this for now, as presumably in future the existing documentation pages will include content specific to `<amp-story-animation>`, but for now this just makes the documentation look broken otherwise.

* Restructure "optimized images" section

Currently, this page appears badly broken in https://amp.dev/documentation/guides-and-tutorials/develop/style_and_layout/responsive_design/?format=stories

This is due to the fact that the markdown processor does not seem to understand that after the `[sourcecode]` block the unordered list is supposed to continue, and instead closes all currently open containers and starts a fresh unordered list. Avoiding the problem by instead using subheadings and a few tweaks to the prose.

* Attempt to fix missing sidebar on "Make your Web Stories discoverable

Not sure if this fix will do have the full effect, but: currently https://amp.dev/documentation/guides-and-tutorials/optimize-and-measure/discovery_web_stories/?format=stories seems to have a problem/lack the left-hand sidebar. This *may* be due to the unorthodox frontmatter. Normalised this to use the same properties, and order, of another sibling markdown document that *does* work correctly, hoping it's just related to that...

* Missing/insufficient alt (in docs and sample code), redefine 'Fallback'

* Add small reminder about audio and providing same info in text

* Make "tricky" in-page link more accessible

* just having a link text of "star" will be meaningless to users
* the `*` is easy to miss / not understand that it's an actual link, particularly since links aren't underlined/obvious

* Add tip about judicious use of animations

* Tweak alt text

* Add note about content order and templates

* Stray space at line end

* Examples edits

- replicating notes from core documentation (e.g. about judicious use of animations, note about lack of 360 control)
- adding placeholder `alt="..."` to encourage authors to think about alternative text as well

* Add links to prefers-reduced-motion issue

* Address reviews

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>