Update doc about one-time use access token (#6330)
It appears that this section was copied from the Gmail Email Markup
documentation:

https://developers.google.com/gmail/markup/actions/limited-use-access-tokens

The particular mention about one-time use of the token is specific to
the One Click Action feature of Email Markup:

https://developers.google.com/gmail/markup/actions/limited-use-access-tokens

I don't think we should have such restriction for AMP for Email. For
example, an email that allows users to reply to a comment thread inline
should be able to use the same token to perform reply multiple times. We
should let sender decide whether to make the token one-time use only.