Link to fixed version of component reference (#6397)
Currently, some of these links point to versions of the component
reference that don't apply to the email spec, such as amp-list 1.0 which
is banned in the email format. We should point to the version
specifically supported in emails.

We are doing this for components even for components that don't have
more than 1 version, for consistency and being future-proof.