Clarify ESP-delivered email validation
Some developers are apparently confused by the instruction here and
tested the validation by sending the email from, for example, the Gmail
playground. That defeats the purpose of the validation mentioned in this
section, because the playground skips any ESP-specific post-processing
of the AMP email authored by the developer. The purpose is to catch
issues sent by the ESP which may not reproduce when using the
playground.